requirements.txt
asgiref==3.5.2
Django==4.1.3
django-durationwidget==1.0.5
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.6
